Rename _v06 > v06
