--- docs/api/dev/groups.md
@@ -1,4 +1,0 @@
-# OME-Zarr groups
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06
