--- docs/api/dev/hcs.md
@@ -1,9 +1,0 @@
-# High Content Screening
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.hcs
-
-## Plate metadata
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.plate
