--- docs/api/dev/image-label.md
@@ -1,9 +1,0 @@
-# Image Label
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.image_label
-
-## Other types
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.image_label_types
