--- docs/api/dev/image.md
@@ -1,9 +1,0 @@
-# Image
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.image
-
-## Multiscales metadata
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.multiscales
