--- docs/api/dev/labels.md
@@ -1,4 +1,0 @@
-# Labels
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.labels
