--- docs/api/dev/other.md
@@ -1,4 +1,0 @@
-# Other
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.base
