--- docs/api/dev/scene.md
@@ -1,4 +1,0 @@
-# Scene
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.scene
