--- docs/api/dev/well.md
@@ -1,9 +1,0 @@
-# Well
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.well
-
-## Other types
-
-<!-- prettier-ignore -->
-::: ome_zarr_models._v06.well_types
