+++ docs/api/v06/bioformats2raw.md
@@ -1,4 +1,4 @@
 # bioformats2raw
 
 <!-- prettier-ignore -->
-::: ome_zarr_models._v06.bioformats2raw
+::: ome_zarr_models.v06.bioformats2raw
