+++ docs/api/v06/coordinates.md
@@ -1,0 +1,67 @@
+# Coordinates
+
+## Base models
+
+::: ome_zarr_models.v06.coordinate_transforms.CoordinateSystem
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.CoordinateSystemIdentifier
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Axis
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Transform
+options:
+show_root_heading: true
+
+## Coordinate transforms
+
+::: ome_zarr_models.v06.coordinate_transforms.Affine
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Bijection
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.ByDimension
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Coordinates
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Displacements
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Identity
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.MapAxis
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Rotation
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Scale
+options:
+show_root_heading: true
+
+::: ome_zarr_models.v06.coordinate_transforms.Translation
+options:
+show_root_heading: true
+
+## Errors
+
+::: ome_zarr_models.v06.coordinate_transforms.NoAffineError
+options:
+show_root_heading: true
