+++ docs/api/v06/groups.md
@@ -1,0 +1,3 @@
+# OME-Zarr groups
+
+::: ome_zarr_models.v06
