+++ docs/api/v06/image-label.md
@@ -1,0 +1,7 @@
+# Image Label
+
+::: ome_zarr_models.v06.image_label
+
+## Other types
+
+::: ome_zarr_models.v06.image_label_types
