+++ docs/api/v06/image.md
@@ -1,0 +1,7 @@
+# Image
+
+::: ome_zarr_models.v06.image
+
+## Multiscales metadata
+
+::: ome_zarr_models.v06.multiscales
