+++ docs/api/v06/labels.md
@@ -1,0 +1,3 @@
+# Labels
+
+::: ome_zarr_models.v06.labels
