+++ docs/api/v06/other.md
@@ -1,0 +1,3 @@
+# Other
+
+::: ome_zarr_models.v06.base
