+++ docs/api/v06/scene.md
@@ -1,0 +1,3 @@
+# Scene
+
+::: ome_zarr_models.v06.scene
