+++ docs/api/v06/well.md
@@ -1,0 +1,7 @@
+# Well
+
+::: ome_zarr_models.v06.well
+
+## Other types
+
+::: ome_zarr_models.v06.well_types
