+++ mkdocs.yml
@@ -116,17 +116,17 @@ nav:
           - Exceptions: api/common/exceptions.md
           - Well: api/common/well.md
       - Dev:
-          - api/dev/groups.md
+          - api/v06/groups.md
           - Metadata classes:
-              - api/dev/hcs.md
-              - api/dev/image.md
-              - api/dev/image-label.md
-              - api/dev/labels.md
-              - api/dev/scene.md
-              - api/dev/well.md
-              - api/dev/coordinates.md
-              - api/dev/bioformats2raw.md
-              - api/dev/other.md
+              - api/v06/hcs.md
+              - api/v06/image.md
+              - api/v06/image-label.md
+              - api/v06/labels.md
+              - api/v06/scene.md
+              - api/v06/well.md
+              - api/v06/coordinates.md
+              - api/v06/bioformats2raw.md
+              - api/v06/other.md
   - Changelog: changelog.md
   - Contributing: contributing.md
