--- src/ome_zarr_models/_v06/__init__.py
@@ -1,9 +1,0 @@
-from ome_zarr_models._v06.bioformats2raw import BioFormats2Raw
-from ome_zarr_models._v06.hcs import HCS
-from ome_zarr_models._v06.image import Image
-from ome_zarr_models._v06.image_label import ImageLabel
-from ome_zarr_models._v06.labels import Labels
-from ome_zarr_models._v06.scene import Scene
-from ome_zarr_models._v06.well import Well
-
-__all__ = ["HCS", "BioFormats2Raw", "Image", "ImageLabel", "Labels", "Scene", "Well"]
