+++ src/ome_zarr_models/v06/base.py
