+++ src/ome_zarr_models/v06/image_label_types.py
