+++ src/ome_zarr_models/v06/scene.py
@@ -7,13 +7,13 @@ from pydantic import BaseModel, Field
 from pydantic_zarr.v3 import GroupSpec
 
 from ome_zarr_models._utils import TransformGraph, _from_zarr_v3
-from ome_zarr_models._v06.base import BaseGroupv06, BaseOMEAttrs, BaseZarrAttrs
-from ome_zarr_models._v06.coordinate_transforms import (
+from ome_zarr_models.v06.base import BaseGroupv06, BaseOMEAttrs, BaseZarrAttrs
+from ome_zarr_models.v06.coordinate_transforms import (
     AnyTransform,
     CoordinateSystem,
     CoordinateSystemIdentifier,
 )
-from ome_zarr_models._v06.image import Image
+from ome_zarr_models.v06.image import Image
 
 
 class SceneAttrs(BaseModel):
