+++ src/ome_zarr_models/v06/well.py
@@ -4,9 +4,9 @@ from typing import Self
 import zarr
 
 from ome_zarr_models._utils import _from_zarr_v3
-from ome_zarr_models._v06.base import BaseGroupv06, BaseOMEAttrs
-from ome_zarr_models._v06.image import Image
-from ome_zarr_models._v06.well_types import WellMeta
+from ome_zarr_models.v06.base import BaseGroupv06, BaseOMEAttrs
+from ome_zarr_models.v06.image import Image
+from ome_zarr_models.v06.well_types import WellMeta
 
 __all__ = ["Well", "WellAttrs"]
