+++ src/ome_zarr_models/v06/well_types.py
