+++ tests/_v06/test_labels.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 from zarr.abc.store import Store
 
-from ome_zarr_models._v06.labels import Labels, LabelsAttrs
+from ome_zarr_models.v06.labels import Labels, LabelsAttrs
 from tests._v06.conftest import json_to_zarr_group
