+++ tests/_v06/test_plate.py
@@ -4,8 +4,8 @@ import pytest
 from pydantic import ValidationError
 from zarr.abc.store import Store
 
-from ome_zarr_models._v06.hcs import HCS
-from ome_zarr_models._v06.plate import Acquisition, Column, Plate, Row, WellInPlate
+from ome_zarr_models.v06.hcs import HCS
+from ome_zarr_models.v06.plate import Acquisition, Column, Plate, Row, WellInPlate
 from tests._v06.conftest import json_to_zarr_group
